Fix: foerderstrecke.lage_id INT statt BIGINT (MySQL FK errno 150)
major_incident.id ist Integer (INT); die FK-Spalte lage_id war BIGINT →
MariaDB lehnt den Fremdschlüssel ab (errno 150, "incorrectly formed") und
Migration 0172 bricht ab. lage_id auf INT angeglichen (Modell + Migration).
Migration 0172 war auf keiner Umgebung erfolgreich angewandt, daher In-Place-Korrektur.

+++ alembic/versions/0172_foerderstrecke_persistenz.py
@@ -25,7 +25,7 @@ def upgrade() -> None:
             status              VARCHAR(20)  NOT NULL DEFAULT 'entwurf',
             objekt_id           BIGINT       NULL,
             incident_id         BIGINT       NULL,
-            lage_id             BIGINT       NULL,
+            lage_id             INT          NULL,
             route_geojson       TEXT         NULL,
             ansaug_json         TEXT         NULL,
             auslass_json        TEXT         NULL,
